requirements.txt
certifi==2020.6.20
cycler==0.10.0
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.3.1
numpy==1.19.5
opencv-python==4.4.0.46
Pillow==7.2.0
protobuf==3.13.0
pyparsing==2.4.7
python-dateutil==2.8.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
tensorboardX==2.1
threadpoolctl==2.1.0
tqdm==4.54.1
